% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effect.R
\name{add_effect}
\alias{add_effect}
\title{add_effect
X is a matrix (likely made by generate_initial_design)
effect is an effect represented by string of letters (e.g. "AB" or "ADE")
OR effect is a an integer}
\usage{
add_effect(X, effect)
}
\description{
add_effect
X is a matrix (likely made by generate_initial_design)
effect is an effect represented by string of letters (e.g. "AB" or "ADE")
OR effect is a an integer
}
