% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_theme.R
\name{dark_theme}
\alias{dark_theme}
\title{Adds a dark theme to a ggplot object with sensible font sizes
and styles}
\usage{
dark_theme()
}
\description{
Adds a dark theme to a ggplot object with sensible font sizes
and styles
}
