% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_design.R
\name{generate_initial_design}
\alias{generate_initial_design}
\title{generate_initial_deignCreates an X matrix with p columns. Each column is composed of 1s and -1s.

The patterns of + and - are different in each column. For example:
Col 1: + - + - + -...
Col 2: + + - - + +...
p is the number of factors and is only required (accepted) input.}
\usage{
generate_initial_design(p)
}
\description{
generate_initial_deignCreates an X matrix with p columns. Each column is composed of 1s and -1s.

The patterns of + and - are different in each column. For example:
Col 1: + - + - + -...
Col 2: + + - - + +...
p is the number of factors and is only required (accepted) input.
}
