% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_confounded.R
\name{get_confounded}
\alias{get_confounded}
\title{I is identifying relation (I = "ABCD")
Returns all confounded variables as data.frame}
\usage{
get_confounded(I)
}
\description{
I is identifying relation (I = "ABCD")
Returns all confounded variables as data.frame
}
