% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_theme.R
\name{light_theme}
\alias{light_theme}
\title{Adds a light theme to a ggplot object with sensible font sizes
and styles}
\usage{
light_theme()
}
\description{
Adds a light theme to a ggplot object with sensible font sizes
and styles
}
