% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_leverage.R
\name{plot_leverage}
\alias{plot_leverage}
\title{plot_leverage
Plots the leverage points with a horizontal line corresponding to a decision
boundary recommendation.}
\usage{
plot_leverage(model)
}
\description{
plot_leverage
Plots the leverage points with a horizontal line corresponding to a decision
boundary recommendation.
}
