% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCs.R
\name{plot_PCs}
\alias{plot_PCs}
\title{Plot principal components}
\usage{
plot_PCs(df, color_by_str)
}
\arguments{
\item{df}{a dataframe with at least three columns--PC1, PC2, color_var}

\item{color_by_str}{a string specifying how to color the points}
}
\value{
a ggplot object
}
\description{
Plot principal components
}
