% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_st_residuals.R
\name{plot_st_residuals}
\alias{plot_st_residuals}
\title{plot_st_residuals
Plots standardized and studentized residuals on one plot.}
\usage{
plot_st_residuals(model)
}
\description{
plot_st_residuals
Plots standardized and studentized residuals on one plot.
}
