% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{qqplot
Returns a ggplot object that can be costomized using grammar of graphics
instead of R's janky base graphics}
\usage{
qqplot(x)
}
\description{
qqplot
Returns a ggplot object that can be costomized using grammar of graphics
instead of R's janky base graphics
}
