% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{p is a plot (usually made by ggplot() call)
ofile is a string with relative path and name, must end in .png}
\usage{
save_plot(p, ofile)
}
\description{
p is a plot (usually made by ggplot() call)
ofile is a string with relative path and name, must end in .png
}
